+++ fractal/components/loading-spinners/LoadingSpinner/LoadingSpinner.config.js
@@ -1,0 +1,14 @@
+module.exports = {
+    "title": "Loading Spinner",
+    "status": "ready",
+    "collated": true,
+    "preview": "@loading-spinner-preview",
+    "context": {
+        "text": "Please wait..."
+    },
+    "variants": [
+        {"name": "loading-spinner-overlay", "context": {
+            "overlay": true
+        }}
+    ]
+};